+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Next
-* Fixed search filter to include notes
+* Fixed loadout search filter to include notes
 
 ## 7.8.2 <span class="changelog-date">(2022-03-14)</span>